==> Code/4 Preprocessing (V2+V3).Rmd <==
---
title: "R Notebook"
output: html_notebook
---

#-------------------------------------------------------------------------------
# 1.0. Data import
#-------------------------------------------------------------------------------

```{r}
library(readr)
df <- read_csv("Datasets/3_V1_processed.csv", 
    col_types = cols(id = col_character(), 
        end_timestamp_previous_ride = col_datetime(format = "%Y-%m-%d %H:%M:%S"), 
        start_timestamp = col_datetime(format = "%Y-%m-%d %H:%M:%S"), 
        zip_4 = col_double()))
```

# ------------------------------------------------------------------------------
# Split 1
# ------------------------------------------------------------------------------

```{r}
library(dplyr)
library(lubridate)
```

```{r}
df1 <- select(df,-all_of(c("temperature","sunshine","humidity","precipitation"))) 
#select(df, all_of(c("id", "end_timestamp_previous_ride", "start_timestamp", "idle_time_in_hours", "hour","month","day", "event_indicator")))
```


```{r}
df1[['start_timestamp']] <- strptime(df1[['start_timestamp']],
                                 format = "%Y-%m-%d %H:%M:%S")
df1[['end_timestamp_previous_ride']] <- strptime(df1[['end_timestamp_previous_ride']],
                                 format = "%Y-%m-%d %H:%M:%S")

df1$datestart <- date(df1$start_timestamp)
df1$dateend <- date(df1$end_timestamp_previous_ride)
```

# How often does a row need to be duplicated?
# (i.e. how many days in the period including start and end day)
```{r}
df1$duplicates <- apply(X = df1[,c("datestart", "dateend")],
                       MARGIN = 1,
                       FUN = function(x)
                         length(
                           seq.Date(
                             from = as.Date(x[2]),
                             to = as.Date(x[1]),
                             by = "day")
                           )
                       )
```

# Duplicate each row 
```{r}
h <- c(df1$duplicates)
new_df1 <- df1[rep(seq_len(nrow(df1)), h), ]
```

# Add counter that represents how many days have to be added to the duplicate time variables
```{r}
new_df1$counter <- 0

for (i in 2:nrow(new_df1)){
    if (new_df1$id[i-1] == new_df1$id[i]){ 
      new_df1$counter[i] <- new_df1$counter[i-1]+1
      }
}
```

# Change day and month
```{r}
new_df1$day <- wday(new_df1$end_timestamp_previous_ride+1*60*60*24*new_df1$counter)
new_df1$month <- month(new_df1$end_timestamp_previous_ride+1*60*60*24*new_df1$counter)
```

# Dropoff minute
```{r}
new_df1$minute <- minute(new_df1$end_timestamp_previous_ride)
new_df1$second <- second(new_df1$end_timestamp_previous_ride)
```

# Add start and stop columns
```{r}
new_df1$start <- NA
new_df1$stop <- NA
```

# Fill start and stop columns
```{r}
new_df1$start[1] <- 0.00
new_df1$stop[1] <- min(new_df1$idle_time_in_hours[1], 24 - new_df1$hour[1] - new_df1$minute[1]/60 - new_df1$second[i]/60/60)
  
  #min((24-new_df1$hour[1]), (new_df1$idle_time_in_hours[1]))

# fill first entry of every observation
for (i in 2:nrow(new_df1)){
  if (new_df1$id[i-1] != new_df1$id[i]){
    new_df1$start[i] <- 0.00
    new_df1$stop[i] <- min(new_df1$idle_time_in_hours[i], 24 - new_df1$hour[i] - new_df1$minute[i]/60 - new_df1$second[i]/60/60)
  }
}

# fill all other entries
for (i in 2:nrow(new_df1)){
  if (new_df1$id[i-1] == new_df1$id[i]){ 
    new_df1$start[i] <- new_df1$stop[i-1]
    new_df1$stop[i] <- new_df1$start[i] + min(24, new_df1$idle_time_in_hours[i] - new_df1$start[i])
    
    }
}
```

# Sanity check
```{r}
subset(new_df1, start>stop)
```

```{r}
new_df1$start <- round(new_df1$start, 2)
new_df1$stop <- round(new_df1$stop, 2)
```

# Import weather data
```{r}
library(readxl)
weather_data <- read_excel("External data/weather_data.xlsx", 
    col_types = c("text", "numeric", "numeric", "numeric", "numeric"))
```

# Create matching columns in both dataframes
```{r}
weather_data[['date']] <- strptime(weather_data[['date']],
                                 format = "%Y-%m-%d")

new_df1$date <- substr(new_df1$end_timestamp_previous_ride + 1*60*60*24*new_df1$counter, 1, 10)

new_df1[['date']] <- strptime(new_df1[['date']], format = "%Y-%m-%d")
```

# Add order to prevent merge from changing sorting
```{r}
new_df1$order  <- 1:nrow(new_df1)
```

# Merge dataframes
```{r}
new_df1 <- merge(new_df1, weather_data, by= "date")
```

# Using order to rechange sorting
```{r}
new_df1 <- new_df1[order(new_df1$order), ]
```

# precipitation from -0.1 to 0.025 (download issue)
```{r}
for (x in 1:nrow(new_df1)){
  if (new_df1$precipitation[x] < 0){
    new_df1$precipitation[x] <- 0.025
  }
}
```

# Add dropoff variable 
```{r}
new_df1$dropoff <- 0

for (i in 1:nrow(df)){
  if (new_df1$counter[i] == 0){ 
    new_df1$dropoff[i] <- 1
    }
}
```

```{r}
df_output1 <- new_df1[,c(2:13, 49:52, 14:39, 46:47, 53)]
```


# Save final dataframe
```{r}
write.csv(df_output1,"Datasets/4_V2_processed.csv", row.names = FALSE)
```


# ------------------------------------------------------------------------------
# Split 2
# ------------------------------------------------------------------------------

```{r}
new_df1$case <- NA

for (i in 1:nrow(new_df1)){
    if (new_df1$duplicates[i] == 1){ 
      new_df1$case[i] <- "both" # day with dropoff and pickup
      }
}

for (i in 1:nrow(new_df1)){
    if (new_df1$stop[i] == new_df1$idle_time_in_hours[i] && new_df1$duplicates[i] != 1){ 
      new_df1$case[i] <- "pickup" # day with pickup, but no dropoff
      }
}

for (i in 1:nrow(new_df1)){
    if (new_df1$counter[i] == 0 && new_df1$stop[i] != new_df1$idle_time_in_hours[i] && is.na(new_df1$case[i])){ 
      new_df1$case[i] <- "dropoff" # day with dropff, but no pickup
      }
}

for (i in 1:nrow(new_df1)){
    if (is.na(new_df1$case[i])){ 
      new_df1$case[i] <- "none" # day without dropoff and pickup
      }
}

```

# Second counter: necessary splits within a day
```{r}
night <- c(0,1,2,3,4,5,6)
morning <- c(7,8,9,10,11)
noon <- c(12,13,14)
afternoon <- c(15,16,17)
evening <- c(18,19,20,21,22,23)
```

```{r}
new_df1$starthour <- hour(new_df1$start_timestamp) #NEW
```


```{r}
new_df1$duplicates2 <- 0

for (i in 1:nrow(new_df1)){
    if (new_df1$case[i] == "none"){ 
      new_df1$duplicates2[i] <- 5 # day without dropoff and pickup
      }
}


for (i in 1:nrow(new_df1)){
  
    if (new_df1$case[i] == "dropoff"){ 
      
      temp <- c(new_df1$hour[i]:23)
      
          if (length(intersect(temp, morning))>0){
            new_df1$duplicates2[i] = new_df1$duplicates2[i] + 1}
      
          if (length(intersect(temp, noon))>0){
            new_df1$duplicates2[i] = new_df1$duplicates2[i] + 1}
      
          if (length(intersect(temp, afternoon))>0){
            new_df1$duplicates2[i] = new_df1$duplicates2[i] + 1}
      
          if (length(intersect(temp, evening))>0){
            new_df1$duplicates2[i] = new_df1$duplicates2[i] + 1}
      
          if (length(intersect(temp, night))>0){
            new_df1$duplicates2[i] = new_df1$duplicates2[i] + 1}          
      }
}
```

```{r}
for (i in 1:nrow(new_df1)){
    if (new_df1$case[i] == "both"){ 
      temp <- c(new_df1$hour[i]:(new_df1$starthour[i])) 
      
          if (length(intersect(temp, morning))>0){
            new_df1$duplicates2[i] = new_df1$duplicates2[i] + 1}
      
          if (length(intersect(temp, noon))>0){
            new_df1$duplicates2[i] = new_df1$duplicates2[i] + 1}
      
          if (length(intersect(temp, afternoon))>0){
            new_df1$duplicates2[i] = new_df1$duplicates2[i] + 1}
      
          if (length(intersect(temp, evening))>0){
            new_df1$duplicates2[i] = new_df1$duplicates2[i] + 1}
      
          if (length(intersect(temp, night))>0){
            new_df1$duplicates2[i] = new_df1$duplicates2[i] + 1}      
      }
}
```

#hour = dropoff hour(between 0 and 23) #starthour = pickup hour (between 0 and 23)
```{r}
for (i in 1:nrow(new_df1)){
    if (new_df1$case[i] == "pickup"){ 
      temp <- c(0:(new_df1$starthour[i]))
      
          if (length(intersect(temp, morning))>0){
            new_df1$duplicates2[i] = new_df1$duplicates2[i] + 1
          }
          if (length(intersect(temp, noon))>0){
            new_df1$duplicates2[i] = new_df1$duplicates2[i] + 1
          }
          if (length(intersect(temp, afternoon))>0){
            new_df1$duplicates2[i] = new_df1$duplicates2[i] + 1
          }
          if (length(intersect(temp, evening))>0){
            new_df1$duplicates2[i] = new_df1$duplicates2[i] + 1
          }
          if (length(intersect(temp, night))>0){
            new_df1$duplicates2[i] = new_df1$duplicates2[i] + 1
          }      
      }
}

```

# Duplicate each row 
```{r}
h <- c(new_df1$duplicates2)
new_df2 <- new_df1[rep(seq_len(nrow(new_df1)), h), ]
```

# Add new start and stop columns + observation counter
```{r}
new_df2$observation <- NA
```

# Fill observation counter
```{r}
new_df2$observation[1] <- 1

for (i in 2:nrow(new_df2)){

  if (new_df2$id[i-1] != new_df2$id[i]){ # wenn neue id
    new_df2$observation[i] <- 1
  }   

}

for (i in 2:nrow(new_df2)){

  if (new_df2$id[i-1] == new_df2$id[i]){ # wenn vorherige id gleich aktueller
    new_df2$observation[i] <- new_df2$observation[i-1] + 1
  }   

}
```

# Add and fill interval
```{r}
new_df2$interval <- NA

for (i in 1:nrow(new_df2)){

  if (new_df2$observation[i] == 1 && new_df2$hour[i] %in% night){ 
    new_df2$interval[i] <- "night"}
  
  if (new_df2$observation[i] == 1 && new_df2$hour[i] %in% morning){ 
    new_df2$interval[i] <- "morning"}  
  
  if (new_df2$observation[i] == 1 && new_df2$hour[i] %in% noon){ 
    new_df2$interval[i] <- "noon"}  
  
  if (new_df2$observation[i] == 1 && new_df2$hour[i] %in% afternoon){ 
    new_df2$interval[i] <- "afternoon"}  
  
  if (new_df2$observation[i] == 1 && new_df2$hour[i] %in% evening){ 
    new_df2$interval[i] <- "evening"}    
}


for (i in 2:nrow(new_df2)){
  
  if(new_df2$interval[i-1] == "night" && is.na(new_df2$interval[i])){
    new_df2$interval[i] <- "morning"}  
  
  if(new_df2$interval[i-1] == "morning" && is.na(new_df2$interval[i])){
    new_df2$interval[i] <- "noon"}  
  
  if(new_df2$interval[i-1] == "noon" && is.na(new_df2$interval[i])){
    new_df2$interval[i] <- "afternoon"} 
  
  if(new_df2$interval[i-1] == "afternoon" && is.na(new_df2$interval[i])){
    new_df2$interval[i] <- "evening"}

  if(new_df2$interval[i-1] == "evening" && is.na(new_df2$interval[i])){
    new_df2$interval[i] <- "night"}
    
}

```

# Fill start and stop columns
```{r}
new_df2$start2 <- NA
new_df2$stop2 <- NA

new_df2$start2[1] <- 0
new_df2$stop2[nrow(new_df2)] <- new_df2$idle_time_in_hours[nrow(new_df2)]

for (i in 2:nrow(new_df2)){ # fill start of first entry
  if (new_df2$id[i-1] != new_df2$id[i]){
    new_df2$start2[i] <- 0
  }
}

for (i in 1:(nrow(new_df2)-1)){ # fill stop of last entry
  if (new_df2$id[i+1] != new_df2$id[i]){ 
    new_df2$stop2[i] <- new_df2$idle_time_in_hours[i] 
  }
}

new_df2$stop2[1] <- 3.48
new_df2$start2[2] <- new_df2$stop2[1]
```

```{r}
# Add stop of first observation and start of second observation in a sequence
for (i in 2:(nrow(new_df2))){ 
  if (new_df2$id[i-1] != new_df2$id[i] && new_df2$id[i+1] == new_df2$id[i]){ # erste observation mit weiteren (all cases except both)
    
    if (new_df2$interval[i] == "morning"){
        new_df2$stop2[i] <- morning[length(morning)] + 1 - new_df2$hour[i] - new_df2$minute[i]/60 - new_df2$second[i]/60/60
        new_df2$start2[i+1] <- new_df2$stop2[i]}
    
    if (new_df2$interval[i] == "noon"){
        new_df2$stop2[i] <-  noon[length(noon)] + 1 - new_df2$hour[i] - new_df2$minute[i]/60 - new_df2$second[i]/60/60
        new_df2$start2[i+1] <- new_df2$stop2[i]}

    if (new_df2$interval[i] == "afternoon"){
        new_df2$stop2[i] <-  afternoon[length(afternoon)] + 1 - new_df2$hour[i] - new_df2$minute[i]/60 - new_df2$second[i]/60/60
        new_df2$start2[i+1] <- new_df2$stop2[i]}    
    
    if (new_df2$interval[i] == "evening"){
        new_df2$stop2[i] <-  evening[length(evening)] + 1 - new_df2$hour[i] - new_df2$minute[i]/60 - new_df2$second[i]/60/60
        new_df2$start2[i+1] <- new_df2$stop2[i]} 
 
    if (new_df2$interval[i] == "night"){
        new_df2$stop2[i] <-  night[length(night)] + 1 - new_df2$hour[i] - new_df2$minute[i]/60 - new_df2$second[i]/60/60
        new_df2$start2[i+1] <- new_df2$stop2[i]}
  }
}
```

```{r}
# Add remaining values
for (i in 2:(nrow(new_df2)-1)){ 
 
  if (!is.na(new_df2$start2[i]) && is.na(new_df2$stop2[i] && new_df2$id[i+1] == new_df2$id[i])){
       
    if (new_df2$interval[i] == "morning"){
        new_df2$stop2[i] <- new_df2$start2[i] + (morning[length(morning)] + 1 - morning[1]) 
        new_df2$start2[i+1] <- new_df2$stop2[i]} 
 
    if (new_df2$interval[i] == "noon"){
        new_df2$stop2[i] <-  new_df2$start2[i] + (noon[length(noon)] + 1 - noon[1]) 
        new_df2$start2[i+1] <- new_df2$stop2[i]}

    if (new_df2$interval[i] == "afternoon"){
        new_df2$stop2[i] <-  new_df2$start2[i] + (afternoon[length(afternoon)] + 1 - afternoon[1]) 
        new_df2$start2[i+1] <- new_df2$stop2[i]}    
    
    if (new_df2$interval[i] == "evening"){
        new_df2$stop2[i] <-  new_df2$start2[i] + (evening[length(evening)] + 1 - evening[1]) 
        new_df2$start2[i+1] <- new_df2$stop2[i]} 
 
    if (new_df2$interval[i] == "night"){
        new_df2$stop2[i] <-  new_df2$start2[i] + (night[length(night)] + 1 - night[1]) 
        new_df2$start2[i+1] <- new_df2$stop2[i]}    
    
 }
}

# CHECK THIS id=654 !!!
# NIGHT 6 hour duration instead of 7
# MORNING 4 hour duration instead of 5
# NOON 2 hour duration instead of 3
# AFTERNOON 2 hour duration instead 2
# EVENING 5 hour duration instead of 6
# NIGHT 6 hour duration instead of 7
```

```{r}
new_df2$start2 <- round(new_df2$start2, 2)
new_df2$stop2 <- round(new_df2$stop2, 2)
```

# Sanity check
```{r}
subset(new_df2, stop2<start2)
```

# Add dropoff variable 
```{r}
new_df2$dropoff <- 0

for (i in 1:nrow(new_df2)){
  if (new_df2$observation[i] == 1){ 
    new_df2$dropoff[i] <- 1
    }
}
```

# Sun = 1, Mon = 2, Tue = 3, Wed = 4, Thu = 5, Fri = 6, Sat = 7
```{r}
df_output2 <- new_df2[,c(2:13, 49:52, 14:39, 59:60, 53, 58)]
```

# Save final dataframe
```{r}
write.csv(df_output2,"Datasets/5_V3_processed.csv", row.names = FALSE)
```
